refactor(routine): DayResponse 분리
