+++ src/main/java/com/yeolsimee/moneysaving/app/routine/repository/RoutineRepositoryCustom.java
@@ -1,7 +1,7 @@
 package com.yeolsimee.moneysaving.app.routine.repository;
 
+import com.yeolsimee.moneysaving.app.routine.dto.dayresponse.DayResponse;
 import com.yeolsimee.moneysaving.app.routine.entity.WeekType;
-import com.yeolsimee.moneysaving.app.routinehistory.dto.DayResponse;
 
 public interface RoutineRepositoryCustom {
 	DayResponse findRoutineDay(Long userId, String routineDay, WeekType weekType);
