+++ src/main/java/com/yeolsimee/moneysaving/app/routine/repository/RoutineRepositoryImpl.java
@@ -1,8 +1,8 @@
 package com.yeolsimee.moneysaving.app.routine.repository;
 
 import com.querydsl.jpa.impl.JPAQueryFactory;
+import com.yeolsimee.moneysaving.app.routine.dto.dayresponse.DayResponse;
 import com.yeolsimee.moneysaving.app.routine.entity.WeekType;
-import com.yeolsimee.moneysaving.app.routinehistory.dto.DayResponse;
 import lombok.RequiredArgsConstructor;
 
 @RequiredArgsConstructor
