+++ src/main/java/com/yeolsimee/moneysaving/app/routine/service/RoutineService.java
@@ -9,10 +9,10 @@ import com.yeolsimee.moneysaving.app.routine.dto.RoutineDaysData;
 import com.yeolsimee.moneysaving.app.routine.dto.RoutineDaysResponse;
 import com.yeolsimee.moneysaving.app.routine.dto.RoutineRequest;
 import com.yeolsimee.moneysaving.app.routine.dto.RoutineResponse;
+import com.yeolsimee.moneysaving.app.routine.dto.dayresponse.DayResponse;
 import com.yeolsimee.moneysaving.app.routine.entity.Routine;
 import com.yeolsimee.moneysaving.app.routine.entity.WeekType;
 import com.yeolsimee.moneysaving.app.routine.repository.RoutineRepository;
-import com.yeolsimee.moneysaving.app.routinehistory.dto.DayResponse;
 import com.yeolsimee.moneysaving.app.routinehistory.entity.RoutineCheckYN;
 import com.yeolsimee.moneysaving.app.routinehistory.repository.RoutineHistoryRepository;
 import com.yeolsimee.moneysaving.app.user.entity.User;
